requirements.txt
python-telegram-bot>=21.4,<22
python-dotenv>=1.0.1,<2
Pillow>=10.0.0,<11
pyzbar>=0.1.9,<0.2
google-cloud-vision>=3.7.0,<4
opencv-python>=4.10.0.84,<5
numpy>=1.26,<3
Jinja2==3.1.4
